[RENAME] splash 파일 패키지 분리
